added checkpoints and metrics
