+++ metrics/dprnn-rawnet/final_metrics_attention.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 12.56465415133039,
+"si_sdr_imp": 12.59639477881958,
+"stoi": 0.89246445614455,
+"stoi_imp": 0.17855734343750154,
+"pesq": 2.7793948007822036,
+"pesq_imp": 1.1768671003977458
+}
