+++ metrics/dprnn-spe-ira/final_metrics_concatenation.json
@@ -1,0 +1,14 @@
+{
+"si_sdr": 13.49016181917737,
+"si_sdr_imp": 13.521902446666557,
+"sdr": 14.654957445339353,
+"sdr_imp": 14.506849948634075,
+"sir": Infinity,
+"sir_imp": NaN,
+"sar": 14.654957445339353,
+"sar_imp": 14.506849948634075,
+"stoi": 0.9031898446174051,
+"stoi_imp": 0.18928273191035672,
+"pesq": 3.036944142897924,
+"pesq_imp": 1.4344164425134658
+}
