+++ metrics/dprnn-spe/final_metrics_FiLM.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 12.974694937163033,
+"si_sdr_imp": 13.006435564652222,
+"stoi": 0.891432404051265,
+"stoi_imp": 0.1775252913442167,
+"pesq": 2.966081113855044,
+"pesq_imp": 1.3635534134705862
+}
