+++ metrics/dprnn-spe/final_metrics_addition.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 11.947222178377212,
+"si_sdr_imp": 11.9789628058664,
+"stoi": 0.8811135478052657,
+"stoi_imp": 0.1672064350982173,
+"pesq": 2.862120491186778,
+"pesq_imp": 1.25959279080232
+}
