+++ metrics/dprnn-spe/final_metrics_attention.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 13.29310807533438,
+"si_sdr_imp": 13.32484870282357,
+"stoi": 0.8981891480704923,
+"stoi_imp": 0.1842820353634439,
+"pesq": 2.986271833340327,
+"pesq_imp": 1.383744132955869
+}
