+++ metrics/dprnn-spe/final_metrics_attention_libri3mix.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 0.4395078070930128,
+"si_sdr_imp": 3.7757407925597315,
+"stoi": 0.6957341430877897,
+"stoi_imp": 0.09920029257842448,
+"pesq": 1.6570093752145767,
+"pesq_imp": 0.2816196215947469
+}
