+++ metrics/dprnn-spe/final_metrics_attention_noisy.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 2.8452560460199625,
+"si_sdr_imp": 4.869278773713316,
+"stoi": 0.7367126455304215,
+"stoi_imp": 0.0901711813084911,
+"pesq": 1.7514105813503265,
+"pesq_imp": 0.3126375257174174
+}
