+++ metrics/dprnn-spe/final_metrics_concatenation.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 13.120177731866017,
+"si_sdr_imp": 13.151918359355205,
+"stoi": 0.8960073882472267,
+"stoi_imp": 0.18210027554017824,
+"pesq": 2.9914539597034455,
+"pesq_imp": 1.3889262593189875
+}
