+++ metrics/dprnn-spe/final_metrics_multiplication.json
@@ -1,0 +1,8 @@
+{
+"si_sdr": 12.789726101937394,
+"si_sdr_imp": 12.821466729426582,
+"stoi": 0.8915695891408459,
+"stoi_imp": 0.1776624764337974,
+"pesq": 2.830472504377365,
+"pesq_imp": 1.2279448039929073
+}
