+++ metrics/dprnn-tasnet/final_metrics.json
@@ -1,0 +1,14 @@
+{
+"si_sdr": 15.763041371359801,
+"si_sdr_imp": 15.762527122886075,
+"sdr": 16.375129876189835,
+"sdr_imp": 16.19589679231582,
+"sir": 25.522865389661824,
+"sir_imp": 25.343630430140067,
+"sar": 17.200580510528912,
+"sar_imp": -50.84073670576193,
+"stoi": 0.9387818713871587,
+"stoi_imp": 0.22584139381695043,
+"pesq": 3.1536912961006163,
+"pesq_imp": 1.5510029299855232
+}
