+++ requirements.txt
@@ -1,8 +1,13 @@
 asteroid==0.7.0
+asteroid_filterbanks==0.4.0
 hydra-core==1.3.2
-numpy==1.24.2
+matplotlib==3.9.0
+numpy==1.26.4
 omegaconf==2.3.0
 pandas==2.2.2
+scipy==1.13.0
+setuptools==68.2.2
 soundfile==0.12.1
-torch==2.0.0
-torchmetrics==0.11.4
+torch==2.2.2
+torchaudio==2.2.2
+wandb==0.16.6
